Added Exceptions to cover multiple scenarios. Senior Exception of FMGBaseException created as well as multiple other exceptions to wrap low-level exceptions and requests exceptions.
Check whether or not a FMG connection is established prior to making a call, if one is not, a FMGValidSessionException is thrown

ValueError and KeyError are caught and wrapped for when a response is not valid. FMGValueError and FMGResponseNotFormedCorrect handle these issues.

FMGConnectionError and FMGConnectTimeout wrap requests.exceptions.ConnectionError and requests.exceptions.ConnectTimeout so the caller does not need to depend on the requests module to capture exceptions

Mandated CONSTANT return from the module where you will receive a tuple with (code, {response data}) where code is the response["status"]["code"] within the JSON response and the {response data} is the response["data"] returned from the system. If no ["data"] key is available in the response, the entire response is returned.

Fixed bug in initial call to determine if workspace mode was set to normal.